# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateContentRequest": {
        "additionalProperties": false,
        "properties": {
          "reuse_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reuse Version Id"
          },
          "target": {
            "anyOf": [
              {
                "enum": [
                  "douyin_video",
                  "xiaohongshu_video",
                  "xiaohongshu_graphic",
                  "wechat_channels_video"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "weak_seed": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Weak Seed",
            "type": "string"
          }
        },
        "required": [
          "weak_seed"
        ],
        "title": "CreateContentRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "diyu_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "可信 cookie 会话决定租户、品牌、发布账号和操作人；客户端不能切换这些作用域。",
    "title": "笛语双应用 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/content": {
      "post": {
        "operationId": "create_content_api_v1_content_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ContentVersionResponse"
                    },
                    {
                      "$ref": "#/components/schemas/GreetingResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ContentQuestionResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ApplicationHandoffResponse"
                    }
                  ],
                  "title": "Response Create Content Api V1 Content Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "缺少或无效的可信会话。"
          },
          "403": {
            "description": "当前可信会话属于另一应用。"
          },
          "422": {
            "description": "业务失败；生成失败时不会产生半成品版本。"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Content"
      }
    },
    "/api/v1/content-versions/{version_id}/save": {
      "post": {
        "operationId": "save_version_api_v1_content_versions__version_id__save_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "default": "douyin_video",
              "enum": [
                "douyin_video",
                "xiaohongshu_video",
                "xiaohongshu_graphic",
                "wechat_channels_video"
              ],
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "缺少或无效的可信会话。"
          },
          "403": {
            "description": "当前可信会话属于另一应用。"
          },
          "422": {
            "description": "业务失败；生成失败时不会产生半成品版本。"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Save Version"
      }
    },
    "/api/v1/display": {
      "post": {
        "operationId": "create_display_api_v1_display_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDisplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DisplayVersionResponse"
                    },
                    {
                      "$ref": "#/components/schemas/DisplayQuestionResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ApplicationHandoffResponse"
                    }
                  ],
                  "title": "Response Create Display Api V1 Display Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "缺少或无效的可信会话。"
          },
          "403": {
            "description": "当前可信会话属于另一应用。"
          },
          "422": {
            "description": "业务失败；生成失败时不会产生半成品版本。"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Display"
      }
    },
    "/api/v1/display-tasks/{task_id}/revisions": {
      "post": {
        "operationId": "revise_display_api_v1_display_tasks__task_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisplayRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DisplayVersionResponse"
                    },
                    {
                      "$ref": "#/components/schemas/DisplayQuestionResponse"
                    }
                  ],
                  "title": "Response Revise Display Api V1 Display Tasks  Task Id  Revisions Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "缺少或无效的可信会话。"
          },
          "403": {
            "description": "当前可信会话属于另一应用。"
          },
          "422": {
            "description": "业务失败；生成失败时不会产生半成品版本。"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Revise Display"
      }
    },
    "/api/v1/display-tasks/{task_id}/versions/{version}": {
      "get": {
        "operationId": "get_display_api_v1_display_tasks__task_id__versions__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "缺少或无效的可信会话。"
          },
          "403": {
            "description": "当前可信会话属于另一应用。"
          },
          "422": {
            "description": "业务失败；生成失败时不会产生半成品版本。"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Display"
      }
    },
    "/api/v1/tasks/{task_id}/revisions": {
      "post": {
        "operationId": "revise_content_api_v1_tasks__task_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "缺少或无效的可信会话。"
          },
          "403": {
            "description": "当前可信会话属于另一应用。"
          },
          "422": {
            "description": "业务失败；生成失败时不会产生半成品版本。"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Revise Content"
      }
    },
    "/api/v1/tasks/{task_id}/versions/{version}": {
      "get": {
        "operationId": "get_version_api_v1_tasks__task_id__versions__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "default": "douyin_video",
              "enum": [
                "douyin_video",
                "xiaohongshu_video",
                "xiaohongshu_graphic",
                "wechat_channels_video"
              ],
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "缺少或无效的可信会话。"
          },
          "403": {
            "description": "当前可信会话属于另一应用。"
          },
          "422": {
            "description": "业务失败；生成失败时不会产生半成品版本。"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Version"
      }
    },
    "/content": {
      "get": {
        "operationId": "content_workbench_content_get",
        "parameters": [
          {
            "in": "query",
            "name": "task",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "in": "query",
            "name": "notice",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notice"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "default": "douyin_video",
              "enum": [
                "douyin_video",
                "xiaohongshu_video",
                "xiaohongshu_graphic",
                "wechat_channels_video"
              ],
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "缺少或无效的可信会话。"
          },
          "403": {
            "description": "当前可信会话属于另一应用。"
          },
          "422": {
            "description": "业务失败；生成失败时不会产生半成品版本。"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Content Workbench"
      }
    },
    "/display": {
      "get": {
        "operationId": "display_workbench_display_get",
        "parameters": [
          {
            "in": "query",
            "name": "task",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "in": "query",
            "name": "notice",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notice"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "缺少或无效的可信会话。"
          },
          "403": {
            "description": "当前可信会话属于另一应用。"
          },
          "422": {
            "description": "业务失败；生成失败时不会产生半成品版本。"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Display Workbench"
      }
    },
    "/ui/display/generate": {
      "post": {
        "operationId": "ui_display_generate_ui_display_generate_post",
        "responses": {
          "303": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Ui Display Generate"
      }
    },
    "/ui/display/revise": {
      "post": {
        "operationId": "ui_display_revise_ui_display_revise_post",
        "responses": {
          "303": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Ui Display Revise"
      }
    },
    "/ui/generate": {
      "post": {
        "operationId": "ui_generate_ui_generate_post",
        "responses": {
          "303": {
            "description": "可信会话中的表单操作完成后重定向回工作台。"
          },
          "401": {
            "description": "缺少或无效的可信会话。"
          },
          "403": {
            "description": "当前可信会话属于另一应用。"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Ui Generate"
      }
    },
...
  }
}